Refactor CSV base class.
This CL moves the common CSV code into a CSVInput class which is then
used by Nunavut and Nova Scotia importers.
